Isolates unit instantiation logic into Instantiator
Instead of having definitions responsible for the instantiation of their
child units, we should introduce another layer here which will be
responsible for creating instances from definitions.  This allows
caching of unit instances into the instantiator where they can be added
as dependencies to future instantiated units.

Here, we introduce the instantiator object which will create unit
instances from their definitions, keeping most of the existing logic of
bailing out if a given defintion is not found via a panic in place.

Reference-counting is utilized for instantiated units since those unit
instances may belong as dependencies to other instances, and will be
invoked from outside of the instantiator where their references live in
a vector.