Prototyping sysunit utility in Rust
I've been using a utility written in POSIX shell that I call `sysunit`
for a long time to maintain system state on my dev machines and images.  It's
worked pretty well, but as more features get added to it it's become
somewhat troublesome to maintain, and bugs frequently appear.

Because a lot of my workflows now rely upon `sysunit` (I use it for
everything from building my personal website to quickly spinning up
development environments for client work) I really want this to be
robust software.  I've decided to re-implement this tool in Rust as it
can provide statically linked binaries for easy installation, should
compile smaller than Go binaries allowing it to run easily on some of
the bitty Linux boards I'd like to use it on, and the safety of Rust
should provide great benefits to the durability of this tool that's
become so critical to my personal workflows.

I also see a potential for this tool to be useful to others, potentially
resulting in some open-source efforts towards its development.  In that
situation, I feel that the Rust community is more likely to contribute
meaningfully to a minimalist tool such as this.

Here, I've quickly prototyped the base behavior of `sysunit`, providing
the minimum I need to begin using it as a replacement for the
`sysunit.sh` shell script I typically reach for.  This implementation is
low-quality, and cuts many corners at the cost of performance.  However,
it should be somewhat robust and is now functional for my basic needs.