Refactors execution to ease introduction of arbitrary targets
All units are currently run against the local machine.
Sysunit's ultimate goal is to allow us to configure any arbitrary
device.  To enable that, we need to allow execution of units via means
other than direct execution.

Here, we begin refactoring to allow us to introduce additional types of
targets.  This is done by abstracting the concept of Executors behind a
trait, and moving the local execution logic into a new Local Executor.
We will now be able to add additional executor types, selected with the
-a (adapter) option.