Adds adapters
To allow running units against multiple kinds of targets (remote hosts,
VMs, docker containers, etc.) we want to way for users to be
able to execute units via some user-defined mechanism.

Here, adapters are introduced to accomplish that.  With the -a flag
provided, a given executable can be invoked for each run of a unit.
Local and SSH adapters are provided to explore this concept.