Idempotent application/rollback
Currently, we're just applying the given operation to units.

Units adhere to a `check` scheme which allows them to report whether
they are currently present on the target so we need only apply or
rollback when they are not in the desired state.

Here, we introduce filtering of units based on their check state and
only execute apply/rollback operations on those which are not yet in the
desired state.