Adds builtin SSH
The built-in shell script based SSH adapter is relatively slow due its
method of invocation via a new process for each run of a unit.  We would
like to provide a built-in adapter that can maintain a persistent
connection for all runs of a unit across the execution of sysunit.

Here, a new built-in adapter is added for SSH based on the libssh2 rust
API.