Adds integration testing
With the growing complexity of this application, regression testing
would be good to have.  As the internal APIs are still very much in
flux, I've chosen to do this only at the integration level for the time
being.  Because this application is primarily oriented towards managing
the state of POSIX linux environments, integration is most easily
accomplished via containers, for which Docker has been chosen over lxd
and other solutions due to the ergonomics afforded by the docker-compose
tooling.

Here, a new docker-compose configuration is provided which runs some
Alpine linux containers.  A new test director is included with scripts
and SSH configurations to allow those linux containers to execute units
against each-other.  Ruby's serverspec tooling is introduced, which is
capable of running commands on both local and remote hosts and verifying
output is correct.  Tests can now be run in these ephemeral environments
where the results of sysunit runs against multiple hosts can be verified
in one go.