Adds a transporter for remote execution
To execute units on a remote host, we need a way for them to be sent
over.  The concept of a transporter is added to the SSH execution target
here, which allows both executable and directory units to be sent across
to the remote host prior to execution.  It includes a cache of
transported unit IDs so that units are sent over only once.